INTELIGENCIA ARTIFICIAL/requirements.txt
# requirements.txt

Flask
python-dotenv
huggingface-hub
flask-cors
